requirements.txt
python-binance
requests
beautifulsoup4
python-dotenv
